Add more complex array example
We've got no way to pass an array to a function.
Should it be possible to use indexing syntax on a pointer?
Does this cause absolute hell if you have a pointer to an array? Probably!

Maybe permit a cast to `[?]u16`, i.e. some foreign array you haven't
allocated yourself? Just saying "I want to treat this arbitrary pointer
as an array, trust me"?